requirements.txt
allure-pytest==2.15.3
pytest==9.0.1
requests==2.32.5